Fix build parallelism in 'pd-externals' module
I made a mistake in commit 4f95916a289aed825fc which caused the build
to run with `make -j`, so it would start a huge number of compile
commands in parallel and eventually abort.

Fixing this caused the 'pd-externals' build to run mostly with `make -j
1`, so it would run one command at a time and go very slowly. I have
patched the Makefile with the necessary '+' prefix for recursive `make`
invocations so that things now work as expected.